move capabilities e2e page objects to more logically consistent location
