+++ experimental/traffic-portal/nightwatch/page_objects/servers/capabilities/capabilityDetails.ts
